tribefire.cortex/tribefire-platform-test/src/tribefire/platform/impl/topology/CartridgeLiveInstancesTest.java
// ============================================================================
// Copyright BRAINTRIBE TECHNOLOGY GMBH, Austria, 2002-2022
// 
// This library is free software; you can redistribute it and/or modify it under the terms of the GNU Lesser General Public
// License as published by the Free Software Foundation; either version 3 of the License, or (at your option) any later version.
// 
// This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
// 
// You should have received a copy of the GNU Lesser General Public License along with this library; See http://www.gnu.org/licenses/.
// ============================================================================
package tribefire.platform.impl.topology;

import org.junit.Assert;
import org.junit.Ignore;
import org.junit.Test;
import org.junit.experimental.categories.Category;

import com.braintribe.model.service.api.InstanceId;
import com.braintribe.testing.category.Slow;
import com.braintribe.testing.category.VerySlow;

/**
 * {@link CartridgeLiveInstances} tests.
 * 
 */
public class CartridgeLiveInstancesTest extends CartridgeLiveInstancesTestBase {

	@Test
	public void testDisabled() {

		CartridgeLiveInstances cli = new CartridgeLiveInstances();
		cli.setEnabled(false);
		cli.postConstruct();

		InstanceId appInstanceId = InstanceId.T.create();
		appInstanceId.setApplicationId("foo");
		appInstanceId.setNodeId("bar");

		cli.accept(null);
		cli.accept(appInstanceId);

		Assert.assertNull(cli.liveInstances());
		Assert.assertNull(cli.liveInstances(appInstanceId));
		Assert.assertNull(cli.liveInstances(null));

	}

	// ==========================================================================//
	// Node is considered dead if latest heartbeat is older than 500 milliseconds
	// ==========================================================================//

	@Test
	public void testShortLived() throws Exception {
		testCartridgeLiveInstances(2000, false, false, false, false, false);
	}

	@Test
	public void testVeryShortLived() throws Exception {
		testCartridgeLiveInstances(500, false, false, false, false, false);
	}

	/* PGA (27.2.2023)
	 * 
	 * Temporarily disabling. These two tests sometimes fail in CI, I wonder if it's just these two, or it is because they are the first tests with
	 * listeners, or just these 2 use-cases have problem.
	 * 
	 * Only way I was able to reproduce the problem locally was to debug, hit a breakpoint in TestListener.onStartup and keep waiting. When the
	 * timeout hits, the error message is the same as the one from CI: */

	// junit.framework.AssertionFailedError: DeadListener for alpha was called 0 times instead of the expected 1 expected:<1> but was:<0>
	// . . . at tribefire.platform.impl.topology.CartridgeLiveInstancesTestBase$TestListener.assertListener(CartridgeLiveInstancesTestBase.java:383)
	// . . . at tribefire.platform.impl.topology.CartridgeLiveInstancesTestBase.testCartridgeLiveInstances(CartridgeLiveInstancesTestBase.java:119)
	// . . . at tribefire.platform.impl.topology.CartridgeLiveInstancesTest.testVeryShortLivedWithListeners(CartridgeLiveInstancesTest.java:62)

	@Category(Slow.class)
	@Test
	@Ignore // FAILS SOMETIMES IN CI
	public void testVeryShortLivedWithListeners() throws Exception {
		testCartridgeLiveInstances(500, true, false, false, false, false);
	}

	@Category(Slow.class)
	@Test
	@Ignore // FAILS SOMETIMES IN CI
	public void testVeryShortLivedWithUnregisteringListeners() throws Exception {
		testCartridgeLiveInstances(500, true, true, false, false, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingApps() throws Exception {
		testCartridgeLiveInstances(500, false, false, true, false, false);
	}

	@Category(VerySlow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithListeners() throws Exception {
		testCartridgeLiveInstances(500, true, false, true, false, false);
	}

	@Category(VerySlow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithUnregisteringListeners() throws Exception {
		testCartridgeLiveInstances(500, true, true, true, false, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, false, false, false, true, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedWithListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, true, false, false, true, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedWithUnregisteringListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, true, true, false, true, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingAppsPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, false, false, true, true, false);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, true, false, true, true, false);
	}

	@Category(VerySlow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithUnregisteringListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(500, true, true, true, true, false);
	}

	@Test
	public void testVeryShortLivedWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, false, false, false, false, true);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedWithListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, true, false, false, false, true);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedWithUnregisteringListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, true, true, false, false, true);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, false, false, true, false, true);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, true, false, true, false, true);
	}

	@Category(Slow.class)
	@Test
	public void testVeryShortLivedRestartingAppsWithUnregisteringListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(500, true, true, true, false, true);
	}

	// ==========================================================================//
	// Node is considered dead if latest heartbeat is older than 100 milliseconds
	// ==========================================================================//

	@Test
	public void testExtremelyShortLived() throws Exception {
		testCartridgeLiveInstances(100, false, false, false, false, false);
	}
	
	@Test
	public void testExtremelyShortLivedWithListeners() throws Exception {
		testCartridgeLiveInstances(100, true, false, false, false, false);
	}
	
	@Test
	public void testExtremelyShortLivedWithUnregisteringListeners() throws Exception {
		testCartridgeLiveInstances(100, true, true, false, false, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingApps() throws Exception {
		testCartridgeLiveInstances(100, false, false, true, false, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithListeners() throws Exception {
		testCartridgeLiveInstances(100, true, false, true, false, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithUnregisteringListeners() throws Exception {
		testCartridgeLiveInstances(100, true, true, true, false, false);
	}

	@Test
	public void testExtremelyShortLivedPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, false, false, false, true, false);
	}
	
	@Test
	public void testExtremelyShortLivedWithListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, true, false, false, true, false);
	}
	
	@Test
	public void testExtremelyShortLivedWithUnregisteringListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, true, true, false, true, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, false, false, true, true, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, true, false, true, true, false);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithUnregisteringListenersPurgingWhenExpired() throws Exception {
		testCartridgeLiveInstances(100, true, true, true, true, false);
	}

	@Test
	public void testExtremelyShortLivedWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, false, false, false, false, true);
	}
	
	@Test
	public void testExtremelyShortLivedWithListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, true, false, false, false, true);
	}
	
	@Test
	public void testExtremelyShortLivedWithUnregisteringListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, true, true, false, false, true);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, false, false, true, false, true);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, true, false, true, false, true);
	}
	
	@Test
	public void testExtremelyShortLivedRestartingAppsWithUnregisteringListenersWithoutCleanup() throws Exception {
		testCartridgeLiveInstances(100, true, true, true, false, true);
	}

}
